requirements.txt
nose2
language_check
MarkovText
nose-parameterized
